# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FixtureTripCreate": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "scenario_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,127}$",
            "title": "Scenario Id",
            "type": "string"
          },
          "source_type": {
            "const": "fixture",
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "scenario_id",
          "display_name"
        ],
        "title": "FixtureTripCreate",
        "type": "object"
      },
...
      "ProblemResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ProblemDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
      "ScenarioSummary": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "catalog_version": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,127}$",
            "title": "Catalog Version",
            "type": "string"
          },
          "currency": {
            "const": "INR",
            "title": "Currency",
            "type": "string"
          },
          "display_timezone": {
            "const": "Asia/Kolkata",
            "title": "Display Timezone",
            "type": "string"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,127}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "truth_label": {
            "title": "Truth Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "catalog_version",
          "display_timezone",
          "currency",
          "truth_label",
          "available"
        ],
        "title": "ScenarioSummary",
        "type": "object"
      },
...
      "TripCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "catalog_version": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,127}$",
            "title": "Catalog Version",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/TripViewSnapshot"
          },
          "trip_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,127}$",
            "title": "Trip Id",
            "type": "string"
          },
          "trip_version": {
            "minimum": 1.0,
            "title": "Trip Version",
            "type": "integer"
          }
        },
        "required": [
          "trip_id",
          "trip_version",
          "catalog_version",
          "snapshot"
        ],
        "title": "TripCreatedResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ResiliTrip API",
    "version": "resilitrip-api-1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "list_scenarios",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  },
                  "title": "Response List Scenarios",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips": {
      "post": {
        "operationId": "create_trip",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "fixture_creation": {
                  "value": {
                    "display_name": "Asha",
                    "scenario_id": "mumbai-goa-v2",
                    "source_type": "fixture"
                  }
                }
              },
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/FixtureTripCreate"
                  },
                  {
                    "$ref": "#/components/schemas/ManualTripCreate"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{trip_id}": {
      "get": {
        "operationId": "get_trip",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/adoptions": {
      "post": {
        "operationId": "adopt_plan",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "f3_adoption": {
                  "value": {
                    "acknowledge_simulation": true,
                    "expected_trip_version": 2,
                    "plan_id": "plan:F3:v2"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/AdoptionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Adopt Plan",
        "tags": [
          "trips"
        ]
      }
    },
...
  }
}